Hopefully fixed the intermittent failure of some tests
It seems that the sorting of substitutions when replacing input symbols and protecting elementary functions names when using implicit multiplication was not consistent from run to run. Hopefully this is now fixed with an adjusted sorting key, essentially we now first sort based on length of string to substitute (maximum munch) and then on the length of the replacement. Previously only the length of the string to be substituted was taken into consideration. The new key does not guarantee a unique ordering in all cases either, but seems to be consistent enough that there is no longer a problem in practice.